+++ Source/searchOptions.py
@@ -288,5 +288,3 @@ def unitTest():
     print('Playback Functions')
     testCase.setPBFunction(2)
     print(testCase.getPBFunction())
-
-#unitTest()