+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+### 4/5/23
+- removed uninstall.py
+- fixed install to always update the searchOptionType table to include all default options
+- added unitTesting section to default.py
+- finished Create functions for searchProfile
+
 ### 4/3/23
 - renamed install.py to searchProfile.py
   - creates required db, tables, and default data required to save search profiles